Revert "test: add profit details test case"
This reverts commit e127956f4e036b7f0a7d12737be89d6611c62a43.

+++ test/AElf.Contracts.Election.Tests/ElectionContractTestHelpers.cs
@@ -89,7 +89,7 @@ public partial class ElectionContractTests : ElectionContractTestBase
     }
 
     private async Task<TransactionResult> VoteToCandidateAsync(ECKeyPair voterKeyPair, string candidatePublicKey,
-        long lockTime, long amount)
+        int lockTime, long amount)
     {
         var electionStub = GetElectionContractTester(voterKeyPair);
         var voteResult = (await electionStub.Vote.SendAsync(new VoteMinerInput
